feature(task 4):
1. Добавила валидацию

+++ src/main/java/ru/profile/service/CompanyService.java
@@ -3,6 +3,7 @@ package ru.profile.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.profile.exception.EntityNotFoundException;
 import ru.profile.model.Company;
 import ru.profile.repository.CompanyRepository;
 
@@ -13,19 +14,25 @@ public class CompanyService {
     private final CompanyRepository companyRepository;
 
     public Company getCompany(Long id) {
-        return companyRepository.getCompanyById(id);
+        return companyRepository.getCompanyById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Компания с id " + id + " не найдена!"));
     }
 
     public Long addCompany(Company company) {
         return companyRepository.save(company).getId();
     }
 
-    public boolean deleteCompany(Long id) {
+    public void deleteCompany(Long id) {
+        if(!companyRepository.existsById(id)){
+            throw new EntityNotFoundException("Компания с id " + id + " не найдена!");
+        }
         companyRepository.deleteById(id);
-        return false;
     }
 
     public Long putCompany(Long id, Company company) {
+        if(!companyRepository.existsById(id)){
+            throw new EntityNotFoundException("компания с id " + id + " не найдена!");
+        }
         company.setId(id);
         return companyRepository.save(company).getId();
     }
